Rename conditional_probabilities module to statistics
